+++ tests/package-tests-client.js
@@ -1,11 +1,8 @@
 // Import Tinytest from the tinytest Meteor package.
 import { Tinytest } from 'meteor/tinytest'
 
-// Import and rename a variable exported by client index.
-import { name as packageNameClient } from 'meteor/template-package'
+// import { name as packageNameClient } from 'meteor/freedombase:web3-login'
 
-// Write your tests here!
-// Here is an example.
-Tinytest.add('template-package - client', function (test) {
-  test.equal(packageNameClient, 'template-package-client')
+Tinytest.add('freedombase:web3-login - client', function (test) {
+  //test.equal(packageNameClient, 'template-package-client')
 })