+++ CHANGELOG.md
@@ -14,9 +14,9 @@
 * Fix user not being logged in after account creation
 
 ## v0.2.1 - 2022/05/10
-* Fix import/export issue
 
 ### Fixes
+* Fix import/export issue
 
 ## v0.2.0 - 2022/05/10